+++ packages/sidecar/lib/test.dart
@@ -1,2 +1,4 @@
-export 'src/test/utilities/expected_lint.dart';
-export 'src/test/utilities/utilities.dart';
+@Deprecated('Depend on package:sidecar_test instead')
+library sidecar.test;
+
+export 'package:sidecar_test/sidecar_test.dart';