+++ packages/sidecar/test/e2e/sidecar_spec_test.dart
@@ -3,9 +3,7 @@ import 'package:intl_lints/intl_lints.dart';
 import 'package:mockito/mockito.dart';
 import 'package:sidecar/sidecar.dart';
 import 'package:sidecar/src/configurations/sidecar_spec/sidecar_spec.dart';
-import 'package:sidecar/src/test/resources/package_resource.dart';
-import 'package:sidecar/src/test/resources/workspace_resource.dart';
-import 'package:sidecar/src/test/utilities/expected_lint.dart';
+import 'package:sidecar_test/sidecar_test.dart';
 import 'package:test/scaffolding.dart';
 
 import '../helpers/example_file_contents.dart';