+++ packages/sidecar_test/lib/sidecar_test.dart
@@ -1,0 +1,6 @@
+export 'src/utilities/utilities.dart';
+export 'src/utilities/expected_lint.dart';
+
+export 'src/resources/analysis_options.dart';
+export 'src/resources/package_resource.dart';
+export 'src/resources/workspace_resource.dart';